gateway/pyproject.toml
[project]
name = "gateway"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aio-pika>=9.5.8",
    "aiosqlite>=0.22.1",
    "db",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "testcontainers[postgres,rabbitmq]>=4.14.1",
    "uvicorn>=0.40.0",
]

[tool.uv.sources]
db = { path = "../libs/db", editable = true }
